Changed authUser. jsp with using jstl added basket, createOrder.jsp
added new servlets

changed index.jsp

+++ dao/src/main/java/DAO/ItemDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface ItemDao {
     Item getItemByName(String itemName);
 
-    Item getItemByPrice(int priceForOne);
+    Item getItemByPrice(long priceForOne);
 
     void save(Item item);
 }
